making changes in code and folder-structure
