+++ Frontend/src/pages/About.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const About = () => {
+  return <div>this is About page</div>;
+};
+
+export default About;
