+++ Frontend/src/pages/Trips.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Trips = () => {
+  return <div>Trips</div>;
+};
+
+export default Trips;
