--- Frontend/src/shared/components/Navigation/MainNavigation.jsx
@@ -1,23 +1,0 @@
-import { Link } from "react-router-dom";
-import NavLinks from "./NavLinks";
-
-import "./MainNavigation.css";
-
-const MainNavigation = () => {
-    return (
-        <header className="main-navigation__header">
-
-            <h1 className="main-navigation__title">
-                <Link to = "/" > Ghumfir </Link>
-            </h1>
-    
-            <nav className="main-navigation__header-nav">
-               <NavLinks />
-            </nav>
-          
-        </header>
-
-    );
-};
-
-export default MainNavigation;
