--- Frontend/src/trip/pages/NewTrip.js
@@ -1,9 +1,0 @@
-// import React from "react";
-
-const NewTrip = () => {
-
-    return <h2>I am New Trip !</h2>
-    
-};
-
-export default NewTrip;
