--- Frontend/src/trip/pages/Trips.jsx
@@ -1,7 +1,0 @@
-// import React from "react";
-
-const Trips = () => {
-    return <h2>I am all the trips ! </h2>;
-};
-
-export default Trips;
